Move _write_report into gitm/safety/audit.py
Relocate the UTF-8 report writer out of the scheduler and into the safety audit
module; loop.py imports it. Kept the _write_report name (not public) to avoid
colliding with gitm.optimizer.report.write_report that loop.py already imports.
